Refactor user insert logic into UserInsertProcessor class
Move two-phase user insert (bulk + enrich) from UserMetadataUpdater
to new UserInsertProcessor class for better separation of concerns.
Update UserMetadataUpdater to delegate to the new class. Add
UserInsertProcessor.cs to project. Improves maintainability,
testability, and code organization.
